--- a/tinygrad/jit.py
+++ b/tinygrad/jit.py
@@ -22,7 +22,6 @@ def get_input_replace(jit_cache: List[JitItem], input_rawbuffers:List[Buffer]) -
     for i,a in enumerate(ji.rawbufs):
       if a in input_rawbuffers:
         input_replace[(j,i)] = input_rawbuffers.index(a)
-  assert len(set(input_replace.values())) == len(input_rawbuffers), "some input tensors not found"
   return input_replace
 def get_jc_idxs_with_updatable_launch_dims(jit_cache: List[JitItem]) -> List[int]:
   return [j for j,ji in enumerate(jit_cache) if isinstance(ji.prg, CompiledASTRunner) and ((ji.prg.global_size and not all_int(tuple(ji.prg.global_size))) or (ji.prg.local_size and not all_int(tuple(ji.prg.local_size))))]
@@ -95,6 +94,7 @@ class TinyJit(Generic[ReturnType]):
         self.jit_cache.extend(jb)
 
       self.input_replace = get_input_replace(self.jit_cache, input_rawbuffers)
+      assert len(set(self.input_replace.values())) == len(input_rawbuffers), "some input tensors not found"
     elif self.cnt == 0:
       # jit ignore
       self.ret = self.fxn(*args, **kwargs)