gpt-2 timings with CUDA:

new:
ran model in 1.79 ms
total 11.74 ms

old:
ran model in 1.67 ms
total 13.05 ms