@geohot 2023-12-06
Can you write a spec of what you want this to do?