@nimlgen 2023-12-07
Yeah that was not ready. Sorry, should have marked as draft.

Basically we want 2 things: 1) split jit into smaller graphs and 2) create graphs even when there are some operations not supported by the graph.  In the current implementation all non-CompiledASTRunner operations are expected to fail the graph creation.

The algo is pretty simple, just iterating through all jit items, combining consecutive items into batches of the maximum allowed size. If a JIT item is a non-CompiledASTRunner, we flush the current batch (which means creating a graph and putting it into the cache) and then add this jit item as is.

This is all implemented in a single pass. I hope it's more or less understandable. I've also added some comments to clarify this concept in the code.